Changed coordinates from int to double

+++ Models/UserModel.cs
@@ -17,10 +17,9 @@ namespace Assign2.Models
         public string City { get; set; }
         public string Province { get; set; }
         public string PostalCode { get; set; }
-        public int Latitude { get; set; }
-        public int Longitude { get; set; }
+        public double Latitude { get; set; }
+        public double Longitude { get; set; }
         public bool IsNaughty { get; set; }
-
         public DateTime DateCreated = DateTime.Now;
     }
 }
